Allow reusing FastAPI app
