+++ src/dstack/_internal/core/models/backends/dstack.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel, Field
+from typing_extensions import Literal
+
+
+class DstackConfigInfo(BaseModel):
+    type: Literal["dstack"] = "dstack"
+
+
+class DstackConfigValues(BaseModel):
+    type: Literal["aws"] = "aws"
